북마크 추가 및 북마크 목록 불러오기 성공

+++ app/src/main/java/com/kunize/stock_market_simulator/etcData/interestFormat.kt
@@ -3,5 +3,5 @@ package com.kunize.stock_market_simulator.etcData
 data class interestFormat (
     var name: String,
     var price: Int,
-    var rate: Double
+    var rate: Float
 )
